Update and rename test.java to MissingRequiredCryptographicStep.java
